requirements.txt
asgiref==3.2.7
autopep8==1.5.3
beautifulsoup4==4.9.1
certifi==2020.4.5.2
cloudinary==1.21.0
dj-database-url==0.5.0
Django==3.0.7
django-bootstrap4==2.0.1
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.1.2
psycopg2==2.8.5
pycodestyle==2.6.0
python-decouple==3.3
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
whitenoise==5.1.0
